Change how IrInstrs are allocated.
Before we had an Array(IrInstr), and IrInstrs stored pointers into this
array when they needed to reference other instructions. This is really
broken as when the Array needs to be grown it reallocates, invalidating
all of those pointers.

Now we pool alloc all IrInstrs and store pointers into the pool in an
Array(IrInstr *) in each block.