Fix all the broken stuff from the relocation handling changes.
This was all totally broken, but the tests still passed. The reason
being that we weren't patching the call to main from _start, so it
remained at 0. This means the call would just jump to the next
instruction, and then we'd return whatever happened to be in eax, which
was 0.

This also exposes the fact that our pass_pointer test doesn't actually
pass, because it has a caller-save register live across a function call
which gets clobbered, and we don't save caller-save registers yet.

This commit also adds support for absolute relocations (R_X86_64_64), so
that we can pass around the address of globals.