Implement #error
This isn't quite correct: if there is a comment in the preprocessing
tokens making up the error, it is included in the error message even
though it should be deleted in an earlier phase. However this is
actually still standards-compliant - 6.10.5.1 states only that the
diagnostic shall "include the specified sequence of preprocessing
tokens", not that it shall *only* include said tokens.