Fix incorrect union member access.
We would access u.pointee_type even if it was an array type, not a
pointer. This happened to work because the first member of u.array is
elem_type, which is what we want in this case. Here we explicitly do
thie conversion with decay_to_pointer instead.

While we're at it, make decay_to_pointer more user-friendly by decaying
arrays and leaving everything else untouched. All the use sites did this
themselves anyway.