First version of enums.
We don't do proper scoping for enum types yet, but we don't do it for
structs or unions either, so this is fine for now. We also don't handle
incomplete enum types properly, but neither of these should be necessary
for bootstrapping.

We also don't handle enum { A = 234} yet, this will be necessary for
bootstrapping.