Fix invalid codegen for 32->64 zero-extension.
This is handled by a MOV between two 32-bit registers. On x64 this
implicitly zeroes the upper 32 bits, and so works as a zero-extension.
However we were moving between a 32 and 64-bit register instead for some
reason, which isn't even a valid instruction. This code path must have
never even been reached before.