Better implementation of varargs.
For the things it supports, this should be compatible with the ABI,
i.e.: able to call into varargs functions compiled by other compilers.
We don't yet pass any arguments in vector registers, so we don't bother
with any of the code for saving them, and always set al to 0.

Also includes changes to the parser - __builtin_va_arg is a new type of
expression. This is necessary because it needs to take a typename as an
argument. You can't just pass sizeof(typename), because different types
with the same size are handled differently. e.g.: int and float.