Rewrite handling of initializers.
Rather than just producing a list of offsets to initialize we instead
produce a nested structure which matches the structure of the
initializer. This is simpler and should make it easy to reuse this code
for handling global initializers, which need to produce constants
instead.

This version also fixes a bug which was hidden because the tests were
wrong.