Split the preprocessor out of the tokeniser.
There are two main reasons for doing this:
* Having the two phases combined is more difficult to reason about and
  harder to maintain.
* There are several preprocessor features that were difficult to
  implement with the old preprocessor. Namely macro-expanding strings
  multiple times. The old implementation of expansion expanded from
  string -> tokens, so this wasn't possible. Changing to string ->
  string would be possible, but it would be messy to implement. We'd
  have to duplicate logic for detecting macros in input, once in the
  macroexpansion code, and once in the regular tokeniser code when we
  encounter identifiers. With the new preprocessor, this is easy to
  implement (and is present in this commit, along with a test).

Our preprocessor roughly implements translation phases 1 to 5, inclusive
(see C99 5.1.1.2). This includes:
* Deleting backslash-newline.
* Stripping comments.
* Stripping (most) insignificant whitespace.
* Executing preprocessor directives and expanding macros.

To keep track of source location in the preprocessed text, we supplement
the text with an Array of "Adjustment"s. These override source location
at particular locations in the preprocessed text. We emit Adjustments
for macros, comments, #includes, and a few more things.

The new implementation produces better SourceLoc info than the old
implementation, so several tests are updated with this better info. In
particular we now associate tokens resulting from macro expansions to
the invocation of the macro that produced them, rather than the
definition of the macro.