Add support for OSO prefix in the assembler.
"OSO" = "Operand Size Override". This is necessary for some 16-bit
instructions, which are encoded the same as the corresponding
32-bit instruction but with this prefix.

We can't simply pretend the prefix is part of the opcode, because it
needs to precede the REX prefix if present. We may have a REX prefix if
we use the 16-bit version of r8-r15.