Support globals that are pointers to other globals.
We do this using "ASM_GLOBAL_REF". This works for now but will need to
be replaced to support more complicated cases. For example, if I have a
global array or struct which contains pointers to other globals. The
global itself is an ASM_GLOBAL_VAR, and hence just contains an array of
bytes. But we need a relocation pointing into the middle of it.

This will probably be implemented by removing ASM_GLOBAL_{VAR,REF}
entirely, and instead having some data structure that represents the
.data section, which contains an array of objects which are either a
literal byte (sequence of bytes maybe?) or a label.