Remove AsmFunction.{prologue,epilogue}.
This is an implementation detail of asm_gen that doesn't need to be
exposed to the assembler. This is the first of several such cleanups,
with the end goal being that an AsmModule is just a set of sections,
each of which has an associated set of labels, and a bunch of
data/instructions.