Don't emit relocations for function calls in the same TU.
The code changed here in asm.c is completely wrong since we never
initialise global->offset to anything. So we just write a completely
incorrect value for the Fixup, but it doesn't matter because it gets
changed at link-time anyway. Now that we write the correct value, we
don't need the relocation.