Rework the "asm" layer, support full global references.
It had a handful of big problems:

1. Duplicated concepts
AsmGlobal, AsmSymbol, and AsmLabel overlapped quite a bit in
functionality, especially AsmSymbol and AsmLabel. There's no real reason
for both of those to exist.

2. Too high level
AsmGlobal doesn't really make sense from an assembly point of view.
There are no objects that contain data, just data (in various forms) and
labels that attach symbols to locations in that data.

3. Contains too much asm_gen implementation detail
AsmFunction for example. asm_gen cares about this information, but it's
all ephemeral information that can be kept locally in asm_gen_function.
The layer down cares nothing for this information, and there's no reason
to store it in the AsmModule.

The new implementation of AsmModule removes AsmGlobal entirely. It
consists of a few sections which contain data, and AsmSymbols which
label locations in those sections. assemble now just does the work of an
assembler, it doesn't convert AsmVars and combine blocks or whatever. A
text assembler could pretty reasonably create an AsmModule and hand it
to assemble.

The new implementation makes it easy to implement global references
inside other globals, which wasn't really possible with the old
implementation. We don't need ASM_GLOBAL_REF, we just emit zeroes and
Fixups, and it doesn't matter that they're inside other objects because
there's just serialized bytes, not objects.

I think it's a testament to how much better the new design is that this
commit adds functionality while being a net negative to LoC.