Move return values into a fresh vreg.
The live range of the result could be quite large, if the result is used
far away. If it crosses a function call boundary, the return value of
that function will clobber RAX, which will assert-fail the register
allocator.

Big thanks to CReduce for reducing a ~20KB repro down to <100B!