Set size to zero on all symbols.
The previous method didn't work - it failed for static functions, for
example. We currently don't have a reliable mechanism for detecting if a
symbol corresponds to the start of a function. We should probably add
symbol types or something to deal with this. For now we just disable
adding symbol sizes, because GDB does better with zeroes than bogus
values.