Add a ret_void at the end of every function.
This way we ensure every block terminates properly, which makes asm_gen
easier, and prevents weird miscompiles for functions missing a return.
We need to do this for compliance because it's only UB to *use* the
value returned from a function that didn't hit a "return". And even if
it was UB without the use, we'd still want to do this.