Store Preds more simply.
Using an intrusive member on AsmSymbol, since if we have a jump to an
instruction it has to have a corresponding AsmSymbol. Don't know why I
didn't think of this earlier, and thought it would have to be on
AsmInstr.

This vastly improves compile times for pathological cases with very
large functions with lots of virtual registers, but the performance is
still terrible.