Cache information about the largest member of working_set.
Rather than recomputing from scratch every time. There are only two
places this can change: at the bottom of the loop if we insert an extra
element that's larger. Or if pc is the largest element when we remove it
from the working set.

In the first case, we can just check if it's larger than the current
largest element, and update if so. In the second case, we still need to
do a search involving highest_set_bit, but we can start from the u64
in the bitset containing the previous largest element.