Fix two bugs with wrong size stores in __builtin_va_start codegen.
First when storing to next_int_reg_offset. We need to ensure we store 4
bytes, hence we need a fixed size immediate.

Second when storing next_stack_arg. The ADD ends up being a one-byte
add, which will actually work as long as it doesn't overflow (since
x86-64 is little-endian), which is dependent on the particular value of
the stack pointer when this happens. To fix, this needs to be an 8-byte
add, and hence we need to do it in a temp vreg instead. This is better
anyway - we do one store instead of two.