Fix incorrect call to pretty_printf.
We should be passing a char*. Passing a String happens to work because
the contents are packed into the next two registers, and "chars" is
first into the list. Hence, we we call va_arg(..., char *), it pulls out
the char * from the first register as expected. If we called va_arg
again it wouldn't work properly.

This doesn't work correctly when compiled by ncc, since ncc bails on
doing the calling convention properly for small structs, and just passes
all structs on the stack.