Fix ir_gen for switch with default not at the end.
Before we'd always insert an unconditional branch to the default case
when we reached it. But the default case might be before other cases, so
we have to just note where it is and insert the branch at the end.

As a bonus, we avoid emitting redundant branches to switch.after when
there's a default case.