Make the vreg creation/assigning API less error-prone.
Having to remember to call append_vreg or assign_vreg was error-prone
for two reasons:
1) You could forget to call it
2) You could call it, but too late - if there's a call to asm_value (for
   example) in between then you can assign the wrong vreg.

The new API eliminates that by combining the call to get the next vreg
number with the append in one function (new_vreg).

In doing so, the old "assign_vreg" doesn't really make sense any more,
since it's just "instr->vreg_number = vreg_number", which is better to
just inline. However, it's useful to have a version which reaches into
an AsmValue, so the "assign_vreg" name is repurposed for this.