Remove the STT_FILE symbol in executables.
This is intended for relocatable object files, to store the name of the
corresponding source file. Not for the name of the executable for
executable ELF files. All this does is make consistency checks more
annoying, by making the executable different if compiled to a file with
a different name.