Install the freestanding headers as part of building ncc
We need to do this now since we started building libc with ncc only.
When building from a clean state, we'll try building libc before these
headers are installed, and ncc will fail to find them.

It makes more sense to do it this way anyway, we just didn't notice the
issue before due to the sequence of how things were done in our
Makefile.