Fix a silly bug in fread.
We simply weren't subtracting the amount we've already read from the
total amount when determining the size of the next read. We got away
with this so far because nothing that uses fread actually cares about
limiting the amount that is read, and the buffers are all large enough
that we don't go out of bounds.

I ran into this on WSL, where there seems to be a difference in how the
read syscall is implemented (or maybe it's a different kernel version or
something). On the native Linux environments I've tested, the kernel
doesn't care if part of the range from [buffer, buffer+size) is not
mapped, as long as it doesn't need to read into the unmapped part.
Whereas on WSL, it immediately checks the entire range and returns
EFAULT if any of it is unmapped.

So for instance if you try to read 100 bytes from a 10 byte file into a
50 byte buffer (where the memory immediately following the buffer is
unmapped), this works fine on Linux, as there are only 10 bytes to read,
which is still within bounds. Whereas WSL will examine the buffer, find
that part of it is unmapped, and fail, despite the fact that there
wasn't enough data in the file to hit that part anyway.