Extend register allocation to float registers
To be honest I don't really like the way this is implemented with two
separate vreg arrays. I think they should just be merged and the
register allocator modified to deal with both in one pass. But this does
at least work, so I'll check it in to have something to return to if
subsequent changes turn out badly.