Simplify setting up the register save area
Before we tried to be clever and save space by making the register save
area variable size, depending on how many registers are already used for
passing normal arguments. This complicates things (especially once we
start saving vector registers too) and isn't really worth it to save a
few bytes.

Instead, keep it a constant size and layout as defined by the ABI, and
just be slightly clever in avoiding saving registers into it when we
know they won't be used.