Refactor SourceLoc representation
Just use a file offset, and reconstruct the line and column number
when needed. This is simpler and more efficient.

Also change the filename to a String, and propagate that type change
throughout the codebase. We still convert back at some points where we
could use a String, but this is good progress.

I would separate these two things out if I could be bothered.